Update 1.1.0
Major update.

Cleaned the framework to be much simpler

The bot can now dynamically accept any command as long as it is a file in ../commands

The bot can now use async functions

The bot now has a %clear command

The announce command was moved to its own file under /commands

Updated the ping command to include actual ping.